Workaround for "garbage after JSON object"...
This patch proposes a (temporary) workardound for when this type of
error arises. It prints out warning messages about the need of updating
the regular expression, pointing out the offending part in the match
according to the hints in issue#27, and additionally tries to
autorecover from the error by eliminating the offending part right
away.

The added dependency on Try::Tiny can be easily eliminated but I'd
suggest to use this module.